# docs/03-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Card Fraud Transaction Management API",
    "description": "API for ingesting fraud decision events and querying transaction data. Uses PostgreSQL with the fraud_gov schema (shared with card-fraud-rule-management).",
    "version": "0.1.0",
    "x-generated-at": "2026-03-01T07:04:17.541261+00:00"
  },
  "paths": {
...
    "/api/v1/decision-events": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Ingest decision event",
        "description": "Ingest a fraud decision event (HTTP endpoint for dev/testing).",
        "operationId": "ingest_decision_event_api_v1_decision_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Event accepted for processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "409": {
            "description": "Conflicting transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "PAN detected (PCI violation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "List transactions",
        "description": "Get a paginated list of transactions with optional filtering.",
        "operationId": "list_transactions_api_v1_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "card_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by card ID",
              "title": "Card Id"
            },
            "description": "Filter by card ID"
          },
          {
            "name": "ip_address",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by IP address",
              "title": "Ip Address"
            },
            "description": "Filter by IP address"
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by device ID",
              "title": "Device Id"
            },
            "description": "Filter by device ID"
          },
          {
            "name": "device_fingerprint_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by device fingerprint hash",
              "title": "Device Fingerprint Hash"
            },
            "description": "Filter by device fingerprint hash"
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by decision (APPROVE/DECLINE)",
              "title": "Decision"
            },
            "description": "Filter by decision (APPROVE/DECLINE)"
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by country code (not persisted)",
              "title": "Country"
            },
            "description": "Filter by country code (not persisted)"
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by merchant ID",
              "title": "Merchant Id"
            },
            "description": "Filter by merchant ID"
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter from date",
              "title": "From Date"
            },
            "description": "Filter from date"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to date",
              "title": "To Date"
            },
            "description": "Filter to date"
          },
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by review status (PENDING/IN_REVIEW/ESCALATED/RESOLVED/CLOSED)",
              "title": "Review Status"
            },
            "description": "Filter by review status (PENDING/IN_REVIEW/ESCALATED/RESOLVED/CLOSED)"
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by risk level (LOW/MEDIUM/HIGH/CRITICAL)",
              "title": "Risk Level"
            },
            "description": "Filter by risk level (LOW/MEDIUM/HIGH/CRITICAL)"
          },
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by case ID",
              "title": "Case Id"
            },
            "description": "Filter by case ID"
          },
          {
            "name": "rule_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by rule ID (matched rules)",
              "title": "Rule Id"
            },
            "description": "Filter by rule ID (matched rules)"
          },
          {
            "name": "ruleset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ruleset ID",
              "title": "Ruleset Id"
            },
            "description": "Filter by ruleset ID"
          },
          {
            "name": "assigned_to_me",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter transactions assigned to current user",
              "default": false,
              "title": "Assigned To Me"
            },
            "description": "Filter transactions assigned to current user"
          },
          {
            "name": "min_amount",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum transaction amount",
              "title": "Min Amount"
            },
            "description": "Minimum transaction amount"
          },
          {
            "name": "max_amount",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum transaction amount",
              "title": "Max Amount"
            },
            "description": "Maximum transaction amount"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor from previous response",
              "title": "Cursor"
            },
            "description": "Pagination cursor from previous response"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get transaction",
        "description": "Get detailed information about a specific transaction.",
        "operationId": "get_transaction_api_v1_transactions__transaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "include_rules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include rule matches",
              "default": true,
              "title": "Include Rules"
            },
            "description": "Include rule matches"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionQueryResult"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Transaction not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}/combined": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get combined transaction view",
        "description": "Get AUTH and MONITORING events (and matches) for a transaction_id.",
        "operationId": "get_transaction_combined_api_v1_transactions__transaction_id__combined_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "include_rules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include rule matches",
              "default": true,
              "title": "Include Rules"
            },
            "description": "Include rule matches"
          }
        ],
        "responses": {
          "200": {
            "description": "Combined transaction view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedTransactionView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Transaction not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}/overview": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get transaction overview",
        "description": "Get combined view with transaction, review, notes, case, and optional matched rules.",
        "operationId": "get_transaction_overview_api_v1_transactions__transaction_id__overview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "include_rules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include matched rules",
              "default": false,
              "title": "Include Rules"
            },
            "description": "Include matched rules"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOverview"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Transaction not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionEventCreate": {
        "properties": {
          "event_version": {
            "type": "string",
            "title": "Event Version",
            "description": "Event schema version",
            "default": "1.0"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "description": "Business transaction ID (idempotency key)"
          },
          "evaluation_type": {
            "$ref": "#/components/schemas/EvaluationType",
            "description": "Evaluation type: AUTH or MONITORING"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At",
            "description": "When the transaction occurred"
          },
          "produced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Produced At",
            "description": "When the decision was produced"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionDetails"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionType"
          },
          "decision_reason": {
            "$ref": "#/components/schemas/DecisionReason"
          },
          "matched_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleMatch"
            },
            "type": "array",
            "title": "Matched Rules"
          },
          "ruleset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Key",
            "description": "Ruleset key for identification"
          },
          "ruleset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Id",
            "description": "Ruleset UUID"
          },
          "ruleset_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Version",
            "description": "Ruleset version number"
          },
          "risk_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Risk level classification"
          },
          "transaction_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Context",
            "description": "Full payload from rule engine with all evaluation context"
          },
          "velocity_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Snapshot",
            "description": "All velocity states at decision time"
          },
          "velocity_results": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Results",
            "description": "Per-rule velocity calculation results for matched rules"
          },
          "engine_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Metadata",
            "description": "Engine mode, processing time, errors, and runtime metadata"
          },
          "raw_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Payload",
            "description": "Allowlist of original fields (PCI-safe)"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "evaluation_type",
          "occurred_at",
          "produced_at",
          "transaction",
          "decision",
          "decision_reason"
        ],
        "title": "DecisionEventCreate",
        "description": "Decision event for ingestion (HTTP POST body)."
      },
      "DecisionEventResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "accepted"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "ingestion_source": {
            "$ref": "#/components/schemas/IngestionSource"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "ingestion_source",
          "ingested_at"
        ],
        "title": "DecisionEventResponse",
        "description": "Response after event ingestion."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}